+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ### vNEXT
 
+- Support `es6` modules and `tree-shaking` ([PR #151](https://github.com/apollostack/angular2-apollo/pull/151))
+
+### v0.8.0
+
+- Made `mutate()` and `query()` methods to return `Observable` instead of `Promise` ([PR #140](https://github.com/apollostack/angular2-apollo/pull/140))
+- Use types of options (for `watchQuery`, `query`, `mutate`) (
+[PR #145](https://github.com/apollostack/angular2-apollo/pull/145),
+[PR #146](https://github.com/apollostack/angular2-apollo/pull/146),
+[PR #148](https://github.com/apollostack/angular2-apollo/pull/148)
+)
+
 ### v0.7.0
 
 - Added support for **Ahead of Time** compilation ([PR #124](https://github.com/apollostack/angular2-apollo/pull/124))